# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rohit Api Documentation",
        "contact": {
            "name": "developer",
            "email": "rohit7305.rk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://localhost/demo-service-yii2-1836-master/api",
            "description": "local server"
        },
        {
            "url": "http://192.168.12.233/demo-service-yii2-1836/api",
            "description": "local server"
        },
        {
            "url": "http://192.168.11.130/demo-service-yii2-1836/api",
            "description": "local server"
        },
        {
            "url": "http://192.168.11.243/demo-service-yii2-1836/api",
            "description": "local server"
        },
        {
            "url": "http://192.168.2.190/demo-service-yii2-1836/api",
            "description": "local server"
        },
        {
            "url": "http://192.168.2.67/demo-service-yii2-1836/api",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/service/list": {
            "get": {
                "tags": [
                    "service"
                ],
                "summary": "",
                "operationId": "service",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "service name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "category id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "service one"
                                        },
                                        "desc": {
                                            "type": "string",
                                            "example": "service description"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "sub_category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "category": {
                                            "type": "object",
                                            "example": "{'id': 1,'name': 'category one','desc': 'category description'}"
                                        },
                                        "sub_category": {
                                            "type": "object",
                                            "example": "{'id': 1,'name': 'sub category one','desc': 'sub category description'}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}